pyproject.toml
[tool.poetry]
name = "role-roll"
version = "0.1.0"
description = ""
authors = ["TheBoringDude <iamcoderx@gmail.com>"]

[tool.poetry.dependencies]
python = "3.7.11"
requests = "^2.26.0"
"discord.py" = "^1.7.3"
watchdog = { extras = ["watchmedo"], version = "^2.1.3" }
aiohttp = "^3.7.4"
redis = "^3.5.3"
cachetools = "^4.2.2"

[tool.poetry.dev-dependencies]
mypy = "^0.910"
flake8 = "^3.9.2"
black = "^21.7b0"
pytest = "^6.2.4"
watchdog = "^2.1.3"
PyYAML = "^5.4.1"
pylint = "^2.9.3"
python-dotenv = "^0.18.0"
types-requests = "^2.25.0"
types-redis = "^3.5.4"
types-cachetools = "^0.1.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
